Use PyObject reference in to_json
In between the git version I was using for development and 0.1.0,
cpython changed the interface for PyTuple::iter to return references to
elements.

Because handling lists and tuples recurses to_json, and because to_json
does not own the tuple, we need to be able to pass the object by
reference.

There's no need to consume the object to_json receives, so we can just
accept references. (This unfortunately breaks the code between PyList
and PyTuple, so we can't simplify it in a macro as easily now.)